requirements.txt
beautifulsoup4==4.11.1
black==22.3.0
click==8.1.3
feedparser==6.0.10
markdownify==0.11.2
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.2
python-slugify==6.1.2
sgmllib3k==1.0.0
six==1.16.0
soupsieve==2.3.2.post1
text-unidecode==1.3
tomli==2.0.1
